+++ config/urls.py
@@ -42,6 +42,11 @@ urlpatterns = [
     url(r'^samplesheets/', include('samplesheets.urls')),
     url(r'^landingzones/', include('landingzones.urls')),
 
+    # Samplesheets configapp URLs
+    url(
+        r'^samplesheets/bih-germline/',
+        include('samplesheets.configapps.bih_germline.urls')),
+
     # django-db-file-storage URLs (needed for admin, obfuscated for users)
     url(r'^xu7in5zs9lylar0n/', include('db_file_storage.urls')),