+++ samplesheets/configapps/bih_germline/plugins.py
@@ -1,4 +1,7 @@
+from projectroles.plugins import get_backend_api
+
 from samplesheets.plugins import SampleSheetConfigPluginPoint
+from samplesheets.utils import get_last_material_index
 
 
 class SampleSheetConfigPlugin(SampleSheetConfigPluginPoint):
@@ -26,11 +29,20 @@ class SampleSheetConfigPlugin(SampleSheetConfigPluginPoint):
     #: Required permission for accessing the plugin
     permission = None
 
-    def get_row_path(self, row):
+    def get_row_path(self, assay, table, row):
         """Return iRODS path for an assay row in a sample sheet. If None,
         display default directory.
+        :param assay: Assay object
+        :param table: List of lists (table returned by SampleSheetTableBuilder)
         :param row: List of dicts (a row returned by SampleSheetTableBuilder)
         :return: String with full iRODS path or None
         """
-        # TODO: Implement this for bih_germline
-        return None
+        irods_backend = get_backend_api('omics_irods')
+
+        if not irods_backend:
+            return None
+
+        # Get index of last material's name column
+        idx = get_last_material_index(table)
+        material_name = row[idx]['value']
+        return irods_backend.get_path(assay) + '/' + material_name